\name{support_function}
\alias{support_function}
\title{Helper Function}
\usage{
support_function()
}
\description{
Prints 'Takes in the dictionary outputted either from the format_from_dataframe or format_from_excel function and returns the data as a formatted data frame according to the original U.S. Copyright Office SR Audit Data online database'.
}
\arguments{
\item{data}{Input the dictionary variable from the format_from_dataframe or format_from_excel function}
}

\examples{

}
