\name{uscoauditlog}
\alias{uscoauditlog}
\title{Reformat U.S. Copyright Office SR Audit Data"}
\usage{
clean_str()
}
\description{
Prints 'Clean U.S. Copyright Office Audit_Data Spreadsheet'.
}
\examples{
str_example = "2*J24*Owner1*L12*LSAI*K10*2*C128*SR_STAT_ID2*N14*Closed2*O16*Open"
clean_str(str_example)
}
