\name{clean_format_all}
\alias{clean_format_all}
\title{Main Function}
\usage{
clean_format_all()
}
\description{
Prints 'Takes in a .xlsx file and returns the formatted data into a new .xlsx and .txt file according to the format from the U.S. Copyright Office SR Audit Data online database.'.
}

\arguments{
\item{excelfile}{Input the original raw SR Audit Data spreadsheet}
}

\examples{

}
